pyproject.toml
[project]
name = "rjg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "debugpy>=1.8.17",
    "maya-stubs>=0.4.1",
    "numpy>=2.3.3",
    "ruff>=0.13.0",
]
[tool.ruff]
line-length = 100

[tool.pyright]
extraPaths = [
    "/usr/autodesk/ApplicationPlugins/ngskintools2/Contents/scripts"
]
typeCheckingMode = "off"
